use C shim to start Android app
This should avoid relying on features of the Android builtin shell,
and so hopefully avoid failures like this one
http://git-annex.branchable.com/design/assistant/blog/day_197__template_haskell/#comment-07f90830f78f6495dcbdf90eb8636129

The C shim sets up busybox, and uses its builtin shell to run runshell.

It's important that busybox be configured with
CONFIG_FEATURE_SH_STANDALONE, so that while runshell is running, it
does not rely on either system utilities, or busybox being already
installed.