copyToRemote should return True when the remote already has the key
This got broken in commit e9238e958877dff9d12a5a0ed396e93931de95ce.
I observed a key that had been copied to a remote, but the location
log was out of date, and due to this bug, git annex transferkey failed
and so the file could not be dropped when it was moved to an archive
directory.