revert addition of Includes: sys/event.h in cabal file
That is very dodgy; it makes *every* C file compiled have that header added
to it. Apparently sys/event.h needs some other header files to be included
on some OS's, and so this leads to compile failures in completely unrelated
places to the code that actually uses sys/event.h

This reverts commit c993d8e7101226c8c14c92a5814cca23602f58a4, which added
this with no rationalle and I must have missed in amoung the other patches
when merging.

Also, Utility/kqueue.c already includes sys/event.h